requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
streamlit>=1.35.0
requests>=2.31.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
schedule>=1.2.0
matplotlib>=3.7.0
Pillow>=10.0.0
scipy>=1.11.0
